+++ returns/context/requires_context_ioresult.py
@@ -15,12 +15,12 @@ from typing_extensions import final
 
 from returns._generated.iterable import iterable
 from returns.context import NoDeps
-from returns.hkt import Kind, dekind
+from returns.interfaces import applicative, bindable, mappable
 from returns.io import IO, IOFailure, IOResult, IOSuccess
 from returns.primitives.container import BaseContainer
+from returns.primitives.hkt import Kind3, dekind
 from returns.primitives.types import Immutable
 from returns.result import Result
-from returns.typeclasses import applicative, functor, monad
 
 if TYPE_CHECKING:
     from returns.context.requires_context import RequiresContext
@@ -42,11 +42,10 @@ _FirstType = TypeVar('_FirstType')
 @final
 class RequiresContextIOResult(
     BaseContainer,
-    Kind['RequiresContextIOResult', _ValueType, _ErrorType, _EnvType],
-    Generic[_ValueType, _ErrorType, _EnvType],
-    functor.Functor[_ValueType],
-    applicative.Applicative[_ValueType],
-    monad.Monad[_ValueType],
+    Kind3['RequiresContextIOResult', _ValueType, _ErrorType, _EnvType],
+    mappable.Mappable3[_ValueType, _ErrorType, _EnvType],
+    bindable.Bindable3[_ValueType, _ErrorType, _EnvType],
+    applicative.Applicative3[_ValueType, _ErrorType, _EnvType],
 ):
     """
     The ``RequiresContextIOResult`` combinator.
@@ -192,7 +191,7 @@ class RequiresContextIOResult(
 
     def apply(
         self,
-        container: Kind[
+        container: Kind3[
             'RequiresContextIOResult',
             Callable[[_ValueType], _NewValueType],
             _ErrorType,
@@ -231,7 +230,7 @@ class RequiresContextIOResult(
         self,
         function: Callable[
             [_ValueType],
-            Kind[
+            Kind3[
                 'RequiresContextIOResult',
                 _NewValueType,
                 _ErrorType,