+++ returns/context/requires_context_result.py
@@ -15,11 +15,11 @@ from typing_extensions import final
 
 from returns._generated.iterable import iterable
 from returns.context import NoDeps
-from returns.hkt import Kind, dekind
+from returns.interfaces import applicative, bindable, mappable
 from returns.primitives.container import BaseContainer
+from returns.primitives.hkt import Kind3, dekind
 from returns.primitives.types import Immutable
 from returns.result import Failure, Result, Success
-from returns.typeclasses import applicative, functor, monad
 
 if TYPE_CHECKING:
     from returns.context.requires_context import RequiresContext
@@ -40,11 +40,10 @@ _FirstType = TypeVar('_FirstType')
 @final
 class RequiresContextResult(
     BaseContainer,
-    Kind['RequiresContextResult', _ValueType, _ErrorType, _EnvType],
-    Generic[_ValueType, _ErrorType, _EnvType],
-    functor.Functor[_ValueType],
-    applicative.Applicative[_ValueType],
-    monad.Monad[_ValueType],
+    Kind3['RequiresContextResult', _ValueType, _ErrorType, _EnvType],
+    mappable.Mappable3[_ValueType, _ErrorType, _EnvType],
+    bindable.Bindable3[_ValueType, _ErrorType, _EnvType],
+    applicative.Applicative3[_ValueType, _ErrorType, _EnvType],
 ):
     """
     The ``RequiresContextResult`` combinator.
@@ -181,7 +180,7 @@ class RequiresContextResult(
 
     def apply(
         self,
-        container: Kind[
+        container: Kind3[
             'RequiresContextResult',
             Callable[[_ValueType], _NewValueType],
             _ErrorType,
@@ -220,7 +219,7 @@ class RequiresContextResult(
         self,
         function: Callable[
             [_ValueType],
-            Kind[
+            Kind3[
                 'RequiresContextResult',
                 _NewValueType,
                 _ErrorType,