+++ returns/methods/__init__.py
@@ -1,0 +1,1 @@
+from returns.methods.bind import bind as bind