+++ returns/pointfree/__init__.py
@@ -1,0 +1,1 @@
+from returns.pointfree.bind import bind as bind